refactor(ic-pos): adopt backend/frontend naming and trim the frontend
Align this example with the convention used across the migrated examples:

- Rename the canisters icpos -> backend and icpos_frontend -> frontend
  (icp.yaml, mops.toml, deploy.sh, test.sh, README, vite.config.ts, and
  the frontend actor/env/binding references).
- Rename the backend sources to backend/app.mo + backend/app.types.mo and
  the interface to backend/backend.did, matching every other example.

Clean up the frontend:

- Drop unused dependencies @dfinity/utils and @radix-ui/react-icons (the
  app uses @icp-sdk/* and lucide-react exclusively).
- Reduce package.json to the minimal convention (name "frontend", no stale
  ckBTC description / kristoferlund repository / bitcoin keywords).
- Remove frontend/components.json (shadcn-CLI config with broken paths, not
  used at build time and kept by no other example).
- Remove the upstream CHANGELOG.md (history of the pre-migration project,
  referencing removed features; the original is credited in the README).

Verified: deploy.sh, test.sh, and the frontend build/serve all pass.
