+++ motoko/ic-pos/frontend/package.json
@@ -1,42 +1,18 @@
 {
-  "name": "ic-pos-frontend",
+  "name": "frontend",
   "private": true,
-  "version": "0.4.0",
   "type": "module",
-  "description": "Accept Bitcoin payments (ckBTC) on the Internet Computer using this simple Point of Sale (POS) application.",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/kristoferlund/ic-pos"
-  },
-  "bugs": "https://github.com/kristoferlund/ic-pos/issues",
-  "homepage": "https://github.com/kristoferlund/ic-pos",
-  "author": "Kristofer Lund <kristofer@fmckl.se>",
-  "keywords": [
-    "pos",
-    "bitcoin",
-    "ckbtc",
-    "dfinity",
-    "ic",
-    "vite",
-    "react",
-    "tailwindcss",
-    "typescript"
-  ],
   "scripts": {
     "dev": "vite",
     "prebuild": "npm i --include=dev",
-    "generate": "icp-bindgen --did-file ../backend/icpos.did --out-dir src/bindings --force",
-    "build": "npm run generate && tsc -b && vite build",
-    "preview": "vite preview"
+    "generate": "icp-bindgen --did-file ../backend/backend.did --out-dir src/bindings --force",
+    "build": "npm run generate && tsc -b && vite build"
   },
   "dependencies": {
-    "@dfinity/utils": "~4.1.0",
-    "@icp-sdk/canisters": "~3.6.0",
     "@hookform/resolvers": "^3.1.1",
     "@icp-sdk/auth": "~7.1.0",
+    "@icp-sdk/canisters": "~3.6.0",
     "@icp-sdk/core": "~5.4.0",
-    "@radix-ui/react-icons": "^1.3.0",
     "@radix-ui/react-label": "^2.0.2",
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-switch": "^1.0.3",
